# Image Processing Improvements

**Key Fixes:**
- 🚫 **No more zoom effect** when loading images of different sizes than the original.
- 👁️ **Keypoints now auto-hide** when outside image bounds.


## Performance (1k images ×10 runs)

| Scenario               | Before  | After   | 
|------------------------|---------|---------|
| Load 128 (disk 256)    | 12.23s  | 12.33s  |
| Load 384 (disk 256)    | 12.54s  | 13.69s  |
| Matching sizes         | No change | 
*Note: Upsampling (slowest case) is rarely needed in practice*

## Visual Comparison (~640x460 → 384x384)

| Before                     | After                             |
|-----------------------|-------------------------|
| <img src="https://github.com/user-attachments/assets/f7b2d250-2d7f-4eaf-a222-c0f6571d6c4d" width="300"> | <img src="https://github.com/user-attachments/assets/e9c4ac8f-1358-4d5c-a21b-469ac89d4997" width="300">|